Add Market Cap API documentation (#33)
Add support for the Market Cap API -

Market Cap API supports returning the values as well as the change in values over 24h, 7d, 30d period.

We have introduced a flag which will toggle wether the percent change is returned. By default the percent change will not be returned as part of the response.